requirements.txt
absl-py==0.8.1
alabaster==0.7.12
altair==4.1.0
argcomplete==1.12.0
asn1crypto==1.2.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autograd==1.3
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
bokeh==2.1.1
celluloid==0.2.0
certifi==2020.12.5
cffi==1.13.0
click==7.1.2
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distro==1.5.0
docutils==0.15.2
entrypoints==0.3
fastprogress==1.0.0
future==0.18.2
gast==0.2.2
gogo-gadget==0.2.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
h5py==2.9.0
idna==2.8
imagecorruptions==1.1.0
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==0.23
invoke==1.4.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.13.3
Jinja2==2.10.3
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-tensorboard==0.2.0
jupyterlab==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
littlemcmc==0.2.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
natsort==7.0.1
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.5
notebook==6.0.2
npm==0.1.1
numpy==1.17.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
opt-einsum==3.3.0
optional-django==0.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pipx==0.15.5.1
plotly==4.9.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.4.2
pygpu==0.7.6
PyJWT==1.7.1
pylint==2.4.4
pymc3==3.7
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyro-api==0.1.1
pyro-ppl==1.1.0
pyrsistent==0.15.5
PySocks==1.7.1
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.22.0
retrying==1.3.3
sampyl-mcmc==0.3
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.3.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
statsmodels==0.10.1
style==1.1.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.9.22
toolz==0.10.0
torch==1.8.1
torch-two-sample==0.1
torchvision==0.9.1
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
uncertainty-metrics==0.0.81
update==0.0.1
urllib3==1.24.2
userpath==1.4.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
